+++ test/wallet-deepclean-test.js
@@ -34,6 +34,11 @@ let bobAcct0Info, bobNames, bobBalance, bobHistory;
 const aliceBlinds = [];
 const bobBlinds = [];
 
+const historyOptions = {
+  limit: 100,
+  reverse: false
+};
+
 async function mineBlocks(n, addr) {
   addr = addr ? addr : new Address().toString('regtest');
   for (let i = 0; i < n; i++) {
@@ -141,12 +146,12 @@ describe('Wallet Deep Clean', function() {
   it('should save wallet data', async () => {
     aliceBalance = await alice.getBalance();
     aliceNames = await alice.getNames();
-    aliceHistory = await alice.getHistory();
+    aliceHistory = await alice.listHistory(-1, historyOptions);
     aliceAcct0Info = await alice.getAccount(0);
 
     bobBalance = await bob.getBalance();
     bobNames = await bob.getNames();
-    bobHistory = await bob.getHistory();
+    bobHistory = await bob.listHistory(-1, historyOptions);
     bobAcct0Info = await bob.getAccount(0);
   });
 
@@ -157,12 +162,12 @@ describe('Wallet Deep Clean', function() {
   it('should have erased wallet data', async () => {
     const aliceBalance2 = await alice.getBalance();
     const aliceNames2 = await alice.getNames();
-    const aliceHistory2 = await alice.getHistory();
+    const aliceHistory2 = await alice.listHistory(-1, historyOptions);
     const aliceAcct0Info2 = await alice.getAccount(0);
 
     const bobBalance2 = await bob.getBalance();
     const bobNames2 = await bob.getNames();
-    const bobHistory2 = await bob.getHistory();
+    const bobHistory2 = await bob.listHistory(-1, historyOptions);
     const bobAcct0Info2 = await bob.getAccount(0);
 
     // Account metadata is fine
@@ -193,12 +198,12 @@ describe('Wallet Deep Clean', function() {
   it('should have recovered wallet data', async () => {
     const aliceBalance2 = await alice.getBalance();
     const aliceNames2 = await alice.getNames();
-    const aliceHistory2 = await alice.getHistory();
+    const aliceHistory2 = await alice.listHistory(-1, historyOptions);
     const aliceAcct0Info2 = await alice.getAccount(0);
 
     const bobBalance2 = await bob.getBalance();
     const bobNames2 = await bob.getNames();
-    const bobHistory2 = await bob.getHistory();
+    const bobHistory2 = await bob.listHistory(-1, historyOptions);
     const bobAcct0Info2 = await bob.getAccount(0);
 
     assert.deepStrictEqual(aliceBalance, aliceBalance2);